RecycleScan/requirements.txt
Jinja2==3.1.4	
MarkupSafe==2.1.5	
PyYAML==6.0.1	
certifi==2024.2.2	
cffi==1.16.0	
charset-normalizer==3.3.2	
colorama==0.4.6	
contourpy==1.2.1	
cryptography==42.0.8	
cycler==0.12.1	
filelock==3.14.0
fonttools==4.53.0	
fsspec==2024.5.0	
idna==3.7	
imutils==0.5.4	
intel-openmp==2021.4.0	
kiwisolver==1.4.5	
matplotlib==3.9.0	
mkl==2021.4.0	
mpmath==1.3.0	
networkx==3.3	
numpy==1.26.4	
opencv-python==4.9.0.80	
oracledb==2.2.1	
packaging==24.0	
pandas==2.2.2	
pillow==10.3.0	
pip==24.0	
psutil==5.9.8	
py-cpuinfo==9.0.0	
pycparser==2.22	
pyparsing==3.1.2	
python-dateutil==2.9.0.post0	
pytz==2024.1	
requests==2.32.3	
scipy==1.13.1	
seaborn==0.13.2	
setuptools==65.5.0	
six==1.16.0	
sympy==1.12.1	
tbb==2021.12.0	
thop==0.1.1-2209072238	
torch==2.3.0	
torchvision==0.18.0	
tqdm==4.66.4	
typing_extensions==4.12.1	
tzdata==2024.1	
ultralytics==8.2.27
urllib3==2.2.1